move Link and its variants to core
