+++ src/core/Link/Link.tsx
@@ -1,22 +1,42 @@
-import React, { Component } from 'react';
+import React, { Component, ReactNode } from 'react';
 import { default as styled } from 'styled-components';
 import { asPropType } from '../../utils/typescript';
-import {
-  Link as CompLink,
-  LinkProps as CompLinkProps,
-} from '../../components/Link/Link';
-import { LinkExternal, LinkExternalProps } from './LinkExternal';
-import { LinkSkip, LinkSkipProps } from './LinkSkip';
+import classnames from 'classnames';
 import { baseStyles } from './Link.baseStyles';
+import { HtmlA, HtmlAProps } from '../../reset';
+
+export const baseClassName = 'fi-link';
 
 type LinkVariant = 'default' | 'external' | 'skip';
-export interface LinkProps extends CompLinkProps {
+export interface LinkProps extends HtmlAProps {
+  /** Link url. Link is not focusable without the href */
+  href: string;
+  /** Custom classname to extend or customize */
+  className?: string;
+  /**
+   * Link element displayed content
+   */
+  children: ReactNode;
   variant?: LinkVariant;
   asProp?: asPropType;
 }
 
+class BaseLink extends Component<LinkProps> {
+  render() {
+    const { className, ...passProps } = this.props;
+    return (
+      <HtmlA
+        {...passProps}
+        className={classnames(className, {
+          [baseClassName]: !className || !className.includes(baseClassName),
+        })}
+      />
+    );
+  }
+}
+
 const StyledLink = styled(({ asProp, ...passProps }: LinkProps) => (
-  <CompLink {...passProps} as={asProp} />
+  <BaseLink {...passProps} as={asProp} />
 ))`
   ${baseStyles};
 `;
@@ -25,20 +45,8 @@ const StyledLink = styled(({ asProp, ...passProps }: LinkProps) => (
  * <i class="semantics" />
  * Used for adding a link
  */
-export class Link extends Component<LinkProps | LinkExternalProps> {
-  static external = (props: LinkExternalProps) => <LinkExternal {...props} />;
-
-  static skip = (props: LinkSkipProps) => <LinkSkip {...props} />;
-
+export class Link extends Component<LinkProps> {
   render() {
-    const { variant, ...passProps } = this.props;
-
-    if (variant === 'external' && 'labelNewWindow' in passProps) {
-      return <LinkExternal {...(passProps as LinkExternalProps)} />;
-    }
-    if (variant === 'skip') {
-      return <LinkSkip {...passProps} />;
-    }
-    return <StyledLink {...passProps} />;
+    return <StyledLink {...this.props} />;
   }
 }
